feat: researcher data access application, front-end only
