+++ src/components/DateRangePicker.tsx
@@ -1,5 +1,5 @@
 // React Imports
-import { useState, forwardRef, SetStateAction, Dispatch, useEffect } from 'react'
+import { forwardRef, SetStateAction, Dispatch, useEffect } from 'react'
 
 // MUI Imports
 import TextField from '@mui/material/TextField'
