+++ src/components/record-list-table/EnhancedTableHead.tsx
@@ -54,6 +54,7 @@ function EnhancedTableHead(props: EnhancedTableProps) {
             align={headCell.numeric ? 'right' : 'left'}
             padding={headCell.disablePadding ? 'none' : 'normal'}
             sortDirection={orderBy === headCell.id ? order : false}
+            sx={{ width: '100%' }}
           >
             <TableSortLabel
               active={orderBy === headCell.id}
